+++ src/TorrentFile/NameMethods.php
@@ -4,18 +4,25 @@ declare(strict_types=1);
 
 namespace Arokettu\Torrent\TorrentFile;
 
+use Arokettu\Torrent\Exception\RuntimeException;
+use Arokettu\Torrent\MetaVersion;
+
 /**
  * @internal
  */
 trait NameMethods
 {
-    abstract public function getInfoHash(): string;
+    abstract public function getInfoHash(?MetaVersion $version = null, bool $binary = false): ?string;
     abstract public function getName(): ?string;
 
     public function getDisplayName(): string
     {
         $name = $this->getName() ?? '';
-        return $name === '' ? $this->getInfoHash() : $name;
+        if ($name === '') { // unset or empty
+            $name = $this->getInfoHash() ??
+                throw new RuntimeException('Unable to generate a name: both name and hash are missing');
+        }
+        return $name;
     }
 
     public function getFileName(): string